SVG Overlay: add shortcut to navigate to child
Add a shortcut key (]) to navigate to the child of active node. This key
behaves the same as V, i.e., navigates to the largest descendant branch,
with shift+] cyling through sibling branches in order of descending
size.

Also changes the shortcut to navigate to parent from P to [ for
consistency.

Closes acardona/CATMAID#834.