Add {nX} pattern to annotation creation
When entities are annotated, the pattern {nX} with X being a number is
now interpreted in a special way: The whole pattern is replaced by a
number that is incremented for each entity being annotated, starting
from X. If X is omitted, i.e. {n} is used, X is implicitly 1. This pattern can
be used multiple times with different X in an annotation.

Closes acardona/CATMAID#822