Annotations: maintain order of annotated entities in back-end
Especially, when counting patterns ('{nX}') are used, a certain order of
neurons that are annotated is expected. E.g. annotating a list of
skeletons from the selection widget with the pattern '{n}' should
reflect the very same order in annotations with incrementing numbers.

The order of the list passed in as POST parameters is now maintained.