Use a subfolder depdenant coookie name suffix
Django uses the cookie name 'sessionid' for its session information in a cookie
called 'csrftoken' for CSRF information. This is fine if only one instance is
running on a certain domain. If, however, multiple instances run on the same
domain, this naming scheme fails. The same cookie name is then used for both
instances, which leads to the logout in all but one instances if the multiple
instances are opened in the same browser. acardona/CATMAID#412 discusses this
issue.

To fix this, this commit introduces a suffix for the cookie names in use. The
suffix is the MD5 hash of the CATMAID_URL (i.e. the sub-directory CATMAID is
running in, relative to the root).