Neuron search: force 'annotated_with' relation if annotator is given
The annotation search has an 'annotated by' field (only 'by' in the widget),
which was so far only used to check if an annotated entity was related in any
way to another entity and that this relation was created by the user. To only
take annotations into account, the 'annotated_with' relation is now required.