Log front-end errors on backend
Now every time CATMAID's general error handler in the front-end is hit, a log
entry is created on the back-end. By default this information goes to stdout,
but this can easily be changed to write to a file. This is what one would have
to add to settings.py if one wanted to this:

    LOGGING['handlers']['frontendfile'] = {
        'level': 'DEBUG',
        'class': 'logging.FileHandler',
        'filename': '/srv/http/CATMAID/logs/frontend.log',
        'formatter': 'verbose',
    }
    LOGGING['loggers']['catmaid.frontend']['handlers'] = ['frontendfile']

To disable this kind of logging, add:

    LOGGING['loggers']['catmaid.frontend']['handlers'] = ['null']

This log is also only created for errors that are triggered during a session of
logged in user or if the anonymous user has the "can_browse" permission.

Thanks to Andrew Champion for suggesting this feature.