RoiTool: add to user profile and default action
This tool adds the already existing ROI tool to the user selectable
tools. It is turned off by default. The RoiTool now also has a default
action on its confirmation button: creating a new ROI. The
classification editor now overrides this handler with its own (it links
the ROI to a class instance).