Node tags: fix erroneous deletion of label
When the last label link for either connectors or treenodes was deleted,
the label was deleted even if the other node type still had links to it.
The resulted in, e.g., all treenode tags of a certain label being
deleted when deleting the only connector tag for that label.

Related to acardona/CATMAID#869.