Skeleton joining: add name of losing skeleton as optional annotation
When skeletons are joined, one skeleton is merged into the other. The
one merged in will not exist anymore after the join. To remember a
reference to it, the split/merge dialog offers now to add an annotation
with name of the losing skeleton to the winning one. If this name
follows the default pattern ("neuron 12345"), its checkbox is unchecked
by default.

Closes acardona/CATMAID#831.