Add autocomplete to search box of both search tools
This makes the browser remember search terms in the classic search (the
looking glass icon) and the neuron search (the tag plus questionmark
icon) in CATMAID's tracing tool. To make this work, we have to jump
through some hoops, though. The browser only remembers search terms if a
form is actually submitted. Since we overrode this and don't submit the
actual form (but do an AJAX request ourself), a dynamically created
iframe is used to do the submit in there. This is required, because a
submit triggers a reload or redirect. Within the iframe, this doesn't
have any effect on the whole CATMAID page and the search term gets
saved. At least in Chrome this is required, in Firefox it works without
this magic.

Closes acardona/CATMAID#879