Overlay: insert node along invisible active edges
When inserting a node along the nearest edge of the skeleton with
CTRL+ALT, include edges from the active node, even if they are invisible
because they are beyond one section from the current view. Requested by
@dbock.

Performance and simplicity could be greatly improved with a "node
neighbors" API route that returns parent and children of the active node
in a single call without the overhead of branch traversal. However,
since this is currently an infrequent operation, only existing APIs are
used.