Stack: ask for confirmation before closing via button
Override the close button for the Stack window to ask for confirmation,
because closing it closes the project and all open windows. This does
not ask for confirmation when exiting the project via other means, like
the Home link. While overriding the button event directly is inelegant,
overriding it somewhere within the close event chain would require
refactoring window management because it is assumed no window can
cancel closure. This approach is less intrusive.

Closes acardona/CATMAID#314.