Tile layer: fix view culling of tiles when zoomed
Fix viewport culling of tiles at the bottom right of the view in two
ways. First, since stack.{x|y} is the center of view, not the origin,
only half of view{Width|Height} should be added, respectively. Second,
while adding positive panning {xd|yd} helps preload tiles when panning
down or right, it should not be added when panning up or left as it
removes required tiles from the right/bottom.