3D viewer: improve picking of objects
When the ray shot through the mouse click position doesn't hit a
selectable object, more rays are casted. The new rays are on concentric
circles around the original position. In total 10 circles are tried, on
each circle there are 4 * i rays (4 on first, 8 on second, ...) and each
circle has a radius larger by half a pixel (in screen space) than the
one before. This means, in the worst case, if no object is hit, 221 rays
are shot without any result. With 50 bigger neurons loaded, this takes
less than one second.

This should solve acardona/CATMAID#890.