Add ReviewerWhitelist model, controls, and tests
Add a ReviewerWhitelist model to represent which reviewers are trusted
by each user for each project. The whitelist also includes starting
times for each trusted user, indicating that only reviews after that
time for that user should be trusted. On the front-end this will allow
the ability aggregate statistics and analysis only for reviews filtered
by this whitelist.

A semi-RESTful API route for retrieving and updating this whitelist is
included at {project_id}/user/reviewer-whitelist. For simplicity this
route currently operates only in bulk -- updates replace that entire
whitelist. Tests are included for this route.

See acardona/CATMAID#880.