Add key shortcut to delete all tags: Shift+T
Recently, the behavior of tag removal has been changed (e.g. when
pressing Shift+K or Shift+L) to not remove all existing tags, but only
the selected one. With this a means of removing all tags was lost. This
commit adds this functionality again, it is mapped to Shift+T shortcut.
This is constant with all the other tag removing shortcuts.

Fixes acardona/CATMAID#886