Add reviewer whitelist front-end
Add storage (as part of ReviewSystem) and settings for reviewer
whitelists. The whitelist is retrieved automatically whenever a project
is loaded or the session user changes. This trades some intrusion into
initialization for reduced complexity in places where the whitelist is
consumed, since it can be accessed synchronously instead of via
continuation. Since whitelist changes are expected to be infrequent,
there is no registration mechansim (a la NeuronNameService), so
consumers must be refreshed to reflect changes to the whitelist.

See acardona/CATMAID#880.