Add support classes for easier performance testing
The PerformanceTest class is meant to be used from a custom script. An
example for this is given in the documentation of the class. It will
create a new test database that is populated from a template data
base---which is expected to be a valid up-to-date CATMAID database. The
views passed in are then tested against this new database and the timing
is recorded. When finished, the new database is destroyed again. This is
done to ensure no cache has influence on the timing on the next run.