Overlay: add shortcut to select nodes by radius
Shift+Y allows the user to use a simple measurement tool to select
skeletons in a radius around the active node and add them to the last
used selection widget. Ctrl+shift+Y selects skeletons in the same way
but removes them from the selection. Unlike spatial selection in the 3D
widget, these selections work in the tracing layer, so select only nodes
in a disc (not a sphere) and only from those nodes loaded in the
overlay.

See acardona/CATMAID#904.