Move review location caching to tile layer
Caching image data for quick review over slow connections requires reuse
of TileLayer logic, so it should be the responsibility of TileLayer.
Add methods for queued caching of specific tiles as well as utility
wrappers for this that allow caching by location. This corrects the
previous caching logic by using the normal tile layer logic for scale,
position, etc.

Note that this commit depends on tilesForLocation, which is added in
later commits due to history rewrite hell.