Add options to duplicate objects in admin view
For projects, stacks, overlays and data views there is now the option to
duplicate objects from within the admin view. To copy objects without
their relations, there is now a new action in the list view's action
menu. To duplicate an entity *with* its relations, select the object and
use the "save as new" button.