3D viewer: add support for stepwise neuron visibility in animations
This allows to make individual skeletons only after a certain amount of
time the animations is running. Currently, every time one rotation is
over (regardless of continuous or back-and-forth mode), one additional
skeleton is made visible until all can be seen. This works in both the
live view and the export.

See acardona/CATMAID#917