Don't force JavaScript strict mode on all files
The CATMAID.js file is the first file that is loaded. If it now sets strict mode
and all other files are concatenated with Pipeline, code that does not conform
to strict mode won't run anymore. This broke our PhantomJS tests when I
introduced this file recently.

The same problem could show up with other files setting strict mode, but it
seems okay so far. Therfore, strict mode should only be activated in an IIFE in
each file.