Move global "ui" variable and UI constructor into CATMAID namespace
This required small changes in many places. A good side effect is,
that I could fix some whitespace errors and that all the
initializations of "ui" in various constructors are gone. Instead, the
CATMAID.ui object is now initialized in its accessor function. It is defined
when it is accessed the first time and is essentially a Singleton.

Also, the UI is constructor is now part of the CATMAID namespace as well and is
defined in a IIFE.