3D viewer: make options available to space object
Some functions in a Space objects already need the option object and get
it passed in (which is fine). This is however not possible for all
functions, e.g. functions called as a result of a mouse action won't
have easy access to the WebGLApplication object. There is no effort
involved in keeping this reference in sync, because it is the same
object that the WebGLApplication references.