3D viewer: add a method to pick a node by a color pick map
This method renders the all pickable objects (e.g. all spheres and
radius volumes) to an off screen buffer. Then it checks for a certain
position in this buffer (usually mouse coordinates) and if the resulting
color is mapped to a clickable object, the object's ID is returned. This
is a much more precise way to pick objects. However, it also is slightly
slower than casting a ray.