3D viewer: add suppert four different rotation axes
This extends the animation tools to allow rotation in one of the four axes X, Y,
Z and Current Up. There is still a small problem with the camera view doing
small local rotations on its own. This, however, will be fixed in a separate
commit. The rotations them self work.

See acardona/CATMAID#917