Review widget: fix selection of next unreviewed segment
So far the review widget did not select the next unreviewed segment when
the 'E' key was pressed, but the first unreviewed one in the list (i.e.
the longest unreveiwed segment). This is now changed so that the next
unreviewed segment is selected. If no unreviewed segment is found in the
remainder of the table, the search is continued from top.

See acardona/CATMAID#743.