Node overlay: use scaled node size to reduce arrow length
The arrow length between a node and a connector is reduced so that the
arrow head can be seen fully. So far the unscaled node size was used and
it worked okay in most situations. However, for most resolution this
doesn't work and to make the arrows render correctly with multi view
tracing stacks, this change is required.