Make errors accumulate in a single ErrorDialog
When creating an ErrorDialog, check if an existing error div already
exists. If so, append the new error and detail to that dialog, and if
necessary put a header in the dialog indicating it is for several
errors. Set a maximum height on the dialog so the dismissal button is
still visible to the user. When the dismissal button or the dialog "X"
is pressed, destroy the dialog rather than just closing it, so the
original div is removed.

See acardona/CATMAID#950.