3D viewer: add missing submitter function to color map picking
To load a mesh after it has been hidden for color based picking, a
submitter function is required. Not passing this in, caused the error
described in acardona/CATMAID#955.

Unfortunately, we are not able to use the 3D viewers submitter function,
because it is not easily accessible from a Space instance, where the
code for color based picking lives. This should, however, be not so much
of a problem. A new temporary submitter is created for every picking.