Stack: fix tool breaking on blur
For tools other than the tracing tool, switching focus to a non-stack
window or closing all stacks but one caused the tool to break for the
remaining stack. This is because the stack window BLUR handler
unregisters the tool from the stack, but the tool was not re-registered
since project.setFocusedStack is not called if the stack was the
focusedStack (even if it was not the focused window).

Fix this by always calling setFocusedStack for project when a stack
window is focused. Project always attempts to set the tool for the
focused stack. If it is the existing tool for that stack, it is not
registered.

Closes acardona/CATMAID#659.