Review system: go back to first node after looking back
Pressing Q on the first node of a segment (the leafs), brings one to the next
section in the direction the segment would extend if it continued through the
leaf. This is done to let one quickly check if the node is really the leaf or
if the neuron continues. Pressing W again did so far bring one to the second
node of ths segment, not the first. This commit fixes this behavior so that one
is brought back to the first node of the segment.

Closes acardona/CATMAID#855