Add NeuronController to access neuron related back-end API
The neuron controller should be a central place to access neuron and
skeleton related back-end functionality. It currently only provides two
methods: deleting treenodes and deleting neurons. Each function returns
a promise to encapsulate the asynchronous call. In the simplest way,
these can be used like this:

  CATMAID.neuronController.deleteNeuron(neuronID)
      .then(function(json) {
        // Do something
      });

So this is basically an alternative interface to callback functionality.
It makes queuing of multiple requests easy and allows for easy error
handling.

There is one instance of the neuron controller created as part of the
CATMAID namespace. As can be seen from the example above, it can be
accessed through CATMAID.neuronController. Multiple events are provided
with which on can register. Currently, there is an event that notifies
about a deleted skeleton and there is one that notifies about a changed
skeleton (when e.g. a node is deleted).