Submitter: fix promise interface
The promise support I introduced in 00e81899 was not working properly.
Its problem was that it expected promises to be executed already, hence
they were not queued properly. This is now changed. Promises can now be
added as a continuation. And if it turns a continuation is a promise,
the submitter waits for its fulfilment, before it marks a request as
completed.