SVG overlay: fix node deletion
Node deletion was recently encapsulated into a promise as part of the
neuron controller. This was done to have one central place where nodes
and neurons are deleted, which sends out events so that other parts of
the code could register to. Due to some problems with making the
submitter work with promises, this had to be rewritten as well.