Connectivity widget: fix CSV export
The CSV export could not deal with percent signs within the individual
table cells. These are passed to a jQuery selector which fails when
parsing them. A new function has been added that will catch this error
and return the plain string instead.

Also the header row fields of the exported table are now quoted.

Closes acardona/CATMAID#954