Connectivity widget: have central connector click handlers
When the number of synapses is clicked in either the incoming or the
outgoing partner table, a connector selection widget opens. So far each
partner had its own handler which could decrease performance for larger
tables. This is now changed to have one handler for each table. Click
events will bubble up to it. Due to performance improvements for larger
table this is also related to acardona/CATMAID#706.