SVG overlay: fix delete/update problem
This fixes the recently reported problem that can occur if a node is
deleted and one schedules a node update before the deletion was
executed. Calling the update node function now results in queuing also
the test if an update is actually needed. This is required, because this
test reads out the node container of the SVG overlay, which might have
been altered before the update request is actually sent. To prevent this
inconsistency, both the test and the update request is now encapsulated
in a promise. The function creating this promise is then queued by the
submitter.

The general problem, though, must have been there before, but we have
apparently been lucky so far.

See acardona/CATMAID#990