SVG overlay: make callback promise use more clear during update
Although it doesn't change the outcome in this particular case (the
submitter acting as a promise), it is more clear to assign the return
value of queuing the callback to the previously created promise variable
which is also returned from the function. It is more in line with how
promises are used in general.