SVG overlay: don't waste memory when updating coordinates in status bar
The coordinates in the status bar are updated with every mouse move,
which can be quite often. Looking at Chromium's memory timeline shows
that constantly moving the mouse in medium speed will use about 2MB/s.
This is a lot and might contribute to acardona/CATMAID#967. Not
constructing an array on the fly and letting map create yet another
array reduces memory consumption to about 1MB/s.