Add a connectivity matrix constructor
A connectivity matrix object can be used to get such a matrix for a set
of row and column skeleton IDs. It will ask the back-end for
connectivity matrix and makes a simple array based matrix data structure
available (3D matrix organized rows first, columns second and a two
element array for incoming and outgoing connections for the respective
cell). Unit test is included.