Window maker: add general widget creation function
This function expectes an object that has a createWidgetConfiguration()
method. Such a method should return an object with at least the field
'contentID' and the function 'createContent'. The former should return a
unique ID of the container that is created for the widget. The latter
will be passed a container with the before defined ID to create content
in.

Optionally, the returned object can also contain the fields 'controlsID'
and the function createControls(). If both exist, a controls container
(or button panel) is created with the specified ID and passed to
createControls() to create the controls.

If a widget instance has a destroy() or a resize() function, they are
bound as handlers to the corresponding widget events.