Set TESTING_ENVIRONMENT setting when running tests
Use a custom test suite runner to create this setting when running
tests. While it is generally best to have testing environments mirror
production environments as closely as possible, it is useful for
downstream extensions (like CATSOP) to be able to function differently
during tests so that Django works correctly with external components.