Skeleton source manager: support creation of unbound source selects
So far a source select (drop down element) could only be created for a
specific skeleton source. The function added allows now also to create a
skeleton select drop down without requiring a skeleton source. Support
for additionally filtered sources (which will not be part of the result
select) is added as well.