Selection table: do not use attr to set properties
jQuery 1.9+ removes the abuse of $().attr to set properties instead of
attributes. Checked is a property, so use $().prop to set it instead of
attr. Likewise value is not a valid attribute for img elements, so was
also being set as a property. Instead set the alt and title text for the
remove img correctly.