Support groups in basic skeleton source
If loadAsGroup() is called on this source, all added skeletons are added
to a group of the given name. The order of skeletons that that are part
of this source remains the same, even if new skeletons are added to a
group. If new skeletons are added as a group and some of them exist
already in this source, they are moved into the group (by default). This
behavior can be switched off, so that existing skeletons won't be added
to the new group.