Connectivity matrix: add grouping support
Skeletons can now be added as a group, for both rows and columns. To
create a new group based on a selected skeleton source, the check box
"As group" has to be checked. A dialog will appear, asking for the new
group name and making sure it is valid. Values of grouped skeletons are
aggregated.

See acardona/CATMAID#918