Connectivity widget: support regex partner filtering
Treat the partner filtering search input as a regular expression when it
begins with "/" and ignore any trailing "/". Otherwise treat it as plain
text and use DataTables' smart search. Creating a custom search input is
necessary as DataTables throws an exception when a partial regex is
entered (e.g., "["), but does not provide a callback to catch this
exception.

Closes acardona/CATMAID#966.