Treenode table: remove ability to edit radius
Editing radii from the treenode table did not provide appropriate cues
to users that it was possible and removed radius annotation from the
context of image data. Since the radius editing tool in the tracing
layer is now robust and quick, remove this feature from the treenode
table.

See acardona/CATMAID#1003.