3D viewer: fix partially wrong adding of missing sections
To add missing slice meshes, the ThreeJS' scene add() method was passed
to a forEach call on the missing slices array. This does not work,
unfortunately, because add() will apply itself on all arguments if more
than one is passed. Now forEach() passes besides the element also an
index an the array to the function. add() now tried to add all three
arguments, which failed for the second and last. This is now fixed and
add() is called explicitly for every mesh. The same is true fro
remove().