3D viewer: fix some add/remove calls in forEach loops
Like explained in b07433a5a9f, ThreeJS' add() and remove() methods will
apply itself on all arguments if there is more than one. This makes it
unsuitable to pass it directly into a forEach call. This commit fixes
the places in 3D viewer that expected add() and remove() to only use the
first argument.