Connectivity matrix widget: add support for cell coloring
This adds all colorbrewer color options plus "No color", "Light red" and
"Light cyan" to the connectivity matrix widget. If a color range is
picked, the colors are distributed between minimum and maximum number of
available connections. For now the minimum is always zero
(alternatively, one could imagine to set this to the synapse threshold).
Colors can be set for pre-synaptic and post-synaptic cells individually.

Colors default to light cyan for post-synaptic connections (row to
column) and light red for pre-synaptic (column to row) connections,
similar to what they represent in the tracing overlay.

See acardona/CATMAID#918