Add dedicated connectivity matrix back-end function
With /<pid>/skeleton/connectivity_matrix one can now retrieve
connectivity counts between two sets of neurons. Returned is a object
that maps source skeleton IDs to a set of target skeleton IDs (both are
part of either input set). The target set is in fact a mapping as well
and contains the individual counts for each connection.