SVG overlay: allow node objects and IDs to be promised
The promiseNode() function can now be used with a node object (like before) or
with only the node ID. If only the node ID (as a string or number) is passed,
the function tries to get it from the node cache. It will throw an error if it
can't find it.