Neuron navigator: handle invalid regex searches gracefully
If a user enters an invalid regular expression into the search fields of
a neuron node or a neuron list node, the search field will turn red and
no update is performed. If the regex is valid for JavaScript, but not
for the back-end, a warning is now displayed instead of an error dialog
and the search field will turn red as well.

Fixes acardona/CATMAID#1008