SVG overlay: trigger skeleton change event after update
After a treenode and a link have been created, a skeleton change event is
triggered. So far this happened right after the async call to the back-end for
creating the link has been made. This commit changes this behavior so that the
event is triggered after the node update caused by a successful link creation is
done.

So far the skeleton change event was triggered right after a link crea