SVG overlay: add node-placement-and-radius-edit mode
If enabled through the settings widget, the radius for a node will be edited
immediately after it has been created. This allows for easier volumetric
reconstruction. In this mode, the radius circle editing tool is used to specify
the radius. No dialog is shown once a radius is selected for a node and it will
only be saved for the new node.

Closes acardona/CATMAID#947