Connectivity widget: fix handler accumulation
Each time tables were re-created in a connectivity widget, an additional
handler was bound to the click event for neuron names in the tables.
This led to many go-to-nearest-node events being fired when clicking on
a name, or more visibly many alerts popping up in case of errors.

Fix this by clearing the event handler. An alternative would be to bind
handlers separately to the upstream and downstream tables (since both
are deleted when re-creating).