Restrict circles of hell to pre- and post-synaptic connections
So far circles of hell did not restrict the types of connections it
looked at. Due to the new abutting connectors and more connector types
to follow, it can't be assumed anymore that there are only synaptic
connections. This commit therefore restrictrs circles of hell to pre-
and post-synaptic connections.