3D viewer: add active node split approximation for virtual nodes
If the active node is virtual and the active node split shading mode is
in use, an approximation of the behavior for real nodes is now used. The
weight of the virtual node's real child is adjusted so that the visual
appearance is similar to real nodes. This is not perfect though: due to
linear interpolation over bigger distances, the split point might be
hard to see.