Doc: add database backup/restore info to ignore edge table
CATMAID's edge table doesn't need to be backed-up, because it can be
regenerated. This can save some space when doing backups. Though
depending on how fast one needs a recovery to happen, it might still be
a good idea to backup the treenode_edge table. The documentation
includes now information on how to ignore it if there is some time to
regenerate this information.