PixiTileLayer: factor out PixiLayer
Factor a basic PixiLayer that creates the renderer and stage context, a
container for display objects in the layer, and handles all filtering
and blending mode responsibilities. Use this as a mixin to
PixiTileLayer, so that it can be reused elsewhere. While this base
object will be tightly coupled with any layer that uses it, because of
the need to directly access rendering contexts, it does deduplicate and
group common functionality.