Tracing tool: unify stack update suspension implementations
Both the multi view tracing branch as well as regular development added
a "do not update" mode for the tracing overlay. This unifies both
implementations so that the node updating is suspended while moving for
all but the active stack. Once the mouse is released, all other stacks
are updated. If the node limit is hit for the active stack, this stack
is also suspended for the rest of the moving action. Small consecutive
moves will also not trigger an update for the active stack if it was
suspended before. Only after a one second wait time, the update is
performed in such a situation.