Tracing tool: support in-/reactivation of bindings for multiple stacks
So far the inactivation and reactivation of mouse binding for the
tracing tool was only stored for one stack. For multiple tracing stacks,
however, the binding cache has to keep track of multiple bindings with
the same name. Therefore, bindings are now stored per stack, ES6 Maps
are used for this, because objects can be used as keys there.