Replace initial data fixture with data migration
The initial data fixture was used to install default data view types and
data views into the CATMAID database. With Django 1.7 initial data
fixtures are not supported anymore and it is suggested that they are
replaced with data migrations. We don't use Django 1.7, yet, but this
commit makes a future upgrade easier.

The data view types are only added if they don't exist already. Data
views are only created if no view with the same type exist already. For
existing instances, both should exist already. With this no existing
view is removed or overridden. Data views and data view types are not
removed in a backward migration, because the default data cannot be
distinguished from user created data.