Fix sequences in data views initialization migration
Data views and data view types are created as part of migration 0001,
but sequences for these tables' PKs are not updated to reflect this.
This causes migration 0046 (eaf6596dfdce7f873abf7925a29e69b26f1eb46b) to
fail when run on the test database. Update migration 0046 to set the
sequence for these PKs to a safe value.