Arbor.js: new function strahlerMatrix
Returns a matrix of child branches as rows and parent branches as columns,
counting at each cell how many occurrences of e.g. a child branch
with Strahler number 2 shoots off a parent branch of Strahler number 3.
Concept by Casey Schneider-Mizell to analyze the structure of an arbor
and compare it to other arbors.