SVG overlay: fix radius edit and selection
Radius edit and selection were broken because of coordinate system
changes from the multi-view tracing merge. The surround circle callbacks
used to return the radius in project units, but they now return stack
units (because the overlay is again in stack space). Have selectRadius
convert the radius to project units as it used to. Also update radius
selection to account for overlay node coordinates being in stack space.