SVG overlay: add function to get a node on an adjacent section
This function allows to get a node ID and location for a node that is either
between two nodes or on the position of the second one. The location has to be
on the next section after the first node (or optionally before the second). This
is useful to respect virtual nodes when moving between child and parent nodes.

See catmaid/CATMAID#1034