SVG overlay: make parent/child navigation respect virtual nodes
Moving to a child or parent of a node will now by default jump to either
the real child or parent if it is on the same or adjacent section. Or it
will create a virtual node that is on the edge to the real child or
parent on the adjacent section. This also works with cycling child nodes.

Fixes catmaid/CATMAID#1034