Allow non-collinear treenode inserts if user has edit permissions
So far the treenode insert API allowed insertion only if the new node
location was right on the edge between parent and child. This constraint
is now loosened a bit: the position can be anywhere as long as the user
has edit permissions on the neuron in question. For users without such
permissions, the collinearity constraint is still enforced.