SVG overlay: create virtual nodes on location sync
As soon as the view is moved around or the section is changed, nodes
that changed location are updated on the back-end. This update is now
extended so that virtual nodes that were moved will be created, just
after all real nodes were updated.

So just like regular nodes, virtual nodes are not instantiated
immediately when moved, but during an implicit sync.

Fixes catmaid/CATMAID#1033