Make tracing setup dialog more friendly
CATMAID displays a setup/update dialog if there are missing classes,
relations or class_instances for a tracing project. Until now, this
dialog suggested the project had to be initialized for tracing to work.
The wording was a bit unfortunate. If there is already some tracing data
available for a project, the dialog now talks about a required updated.
This is more friendly and doesn't bring data loss to mind.

We might want to think about if we want to automate this and get rid of
the dialog or have it only as a backup (which might be useful when
projects are imported).

Fixes catmaid/CATMAID#1035