Slider: allow a minimum value change for stepped moves
Add a parameter to specify a minimum change in value when moving between
steps. This makes behavior when the actual value is very close to a
stepped value less surprising -- instead of moving to a step that may be
imperceptibly different to the user, attempt to move an additional step.

This is also a kludge to avoid strange cases caused by precision errors
where moving to a step invokes a callback that sets the value to one
slightly different than the step value (for example, exponentiation and
log when setting zoom).