Stack viewer: make navigating with project optional
Add a checkbox option whether to couple navigation of a stack viewer
with the other viewers in the project. When checked, the viewer
behaves as before, moving whenever other stacks in the project move and
vice verse. When unchecked, the stack viewer can be moved independently
of all other stack viewers.