SVG overlay: listen to active node changes
So far recoloring and top bar update have been dealt with in the active
node setting function. This caused only the view to update which caused
the change. These reactions to an active node change are now done in a
handler function that listens to the change event. This makes sure that
nodes are also recolored if the active node change was caused by another
widget or another SVG overlay.