Neuron name service: update clients after adding a new node
When a so far unknown neuron is registered with the neuron name service,
its name will be constructed right away. If new clients register for the
very same neuron before the name is ready, they won't get a new name
until they refresh manually. To fix this, all clients are now updated
once the new name is available.