SVG overlay: only update neuron name in top bar once
The first update was triggered in a callback after the initial
registration and update finished. The second update is caused by the
notification of all clients that is done by the name service after this
update was successful. The top bar will now only update as reaction to
the client notification.