SVG overlay: compensate for Z rounding mismatch when querying nodes
When a stack is moved to a project space position its Z coordinates gets
rounded towards the closest section. So far the node query bounding box
spanned the distance between the stack's current section and the next.
This, however, can lead to nodes visible in one stack and invisible in
another. The problem is that a stack's Z position might have been
rounded to a Z below the actual world position. To compensate for this,
the difference to the actual project position is added to the min
corner of the bounding box. This guarantees that the correct project
space position and the complete stack thickness is taken into account.