Tile sources: use slice position instead of viewer
Do not use the stack viewer to generate tile URLs because this (a) may
be incorrect when the tile source's stack is not the viewer's primary
stack and (b) needlessly couples tile URL generation to stack viewer
state. Instead, pass invariant stack pixel coordinates (just z for 3D
stacks) to the tile source.